Moved load function in blueprint
