+++ source/app/blueprints/rest/alerts_routes.py
@@ -17,6 +17,7 @@
 #  Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 import marshmallow
+from marshmallow.exceptions import ValidationError
 from datetime import datetime
 from flask import Blueprint
 from flask import request
@@ -64,6 +65,14 @@ from app.business.errors import BusinessProcessingError
 from app.business.alerts import alerts_create
 
 alerts_rest_blueprint = Blueprint('alerts_rest', __name__)
+
+
+def _load(request_data, **kwargs):
+    try:
+        alert_schema = AlertSchema()
+        return alert_schema.load(request_data, **kwargs)
+    except ValidationError as e:
+        raise BusinessProcessingError('Data error', data=e.messages)
 
 
 @alerts_rest_blueprint.route('/alerts/filter', methods=['GET'])
@@ -201,11 +210,12 @@ def alerts_add_route() -> Response:
     asset_schema = CaseAssetsSchema()
     assets = asset_schema.load(assets_list, many=True, partial=True)
     try:
-        alert = alerts_create(request_data, iocs, assets)
-        if not user_has_client_access(iris_current_user.id, alert.alert_customer_id):
+        alert = _load(request_data)
+        result = alerts_create(alert, iocs, assets)
+        if not user_has_client_access(iris_current_user.id, result.alert_customer_id):
             return response_error('User not entitled to create alerts for the client')
         alert_schema = AlertSchema()
-        return response_success('Alert added', data=alert_schema.dump(alert))
+        return response_success('Alert added', data=alert_schema.dump(result))
 
     except BusinessProcessingError as e:
         return response_error(e.get_message(), data=e.get_data())
